+++ NEWS.md
@@ -1,6 +1,9 @@
+* Added the function, `add_bibtex`
+
 # add2bibtex 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
 * Added README.
 * Updated README.
 * Added Lisence.
+* Added author